moved notification stuff into separate file
