+++ desktop/app.js
@@ -1,5 +1,6 @@
 const { clipboard, nativeImage, ipcRenderer } = require('electron')
 const { getConnectionCode, startReceivingService } = require('./connection')
+const { notifyCopy, notifyFileSaved } = require('./notifications')
 
 const primaryColor = '#25AE88'
 
@@ -85,36 +86,7 @@ function fileReceivedSuccessfully(filepath, filename) {
   }
 
   notifyFileSaved(filename, filepath)
-}
-
-function notify(title, body, icon, opts = {}, cb) {
-  const notifOptions = {
-    body,
-    icon,
-    silent: true,
-    ...opts,
-  }
-
-  const myNotification = new Notification(title, notifOptions)
-  myNotification.onclick = () => {
-    // we can do something when user click file,
-    // for example open the directory, or preview the file
-  }
-}
-
-
-function notifyCopy(data) {
-  const title = `Received text`
-  const body = data
-  const image = `${__dirname}/trayIconTemplate@2x.png`
-  notify(title, body, image)
-}
-
-function notifyFileSaved(filename, filepath) {
-  const title = `New File`
-  const body = `A new file has been saved, ${filename}`
-  const image = `${__dirname}/trayIconTemplate@2x.png`
-  notify(title, body, isImage(filename) ? filepath : image)
+  addFileToHistory(filename, filepath, new Date());
 }
 
 function isImage(filename) {
